+++ benchmark/Benchmarks.hs
@@ -7,7 +7,7 @@ import Data.Text.Time (parseISODateTime)
 import qualified Data.TimeSeries as TS
 
 
-instance NFData TS.Series where
+instance NFData a => NFData (TS.Series a) where
     rnf xs = rnf (TS.toList xs)
 
 
@@ -20,7 +20,7 @@ index n = posixSecondsToUTCTime (fromIntegral n)
 lastIndex :: UTCTime
 lastIndex = index (seriesSize - 1)
 
-bigSeries :: TS.Series
+bigSeries :: TS.Series Double
 bigSeries = TS.tsSeries [1..seriesSize] [1..]
 
 
@@ -30,7 +30,7 @@ main = defaultMain
         [ bench "size"  $ nf TS.size bigSeries
         , bench "valueAt"  $ nf (\xs -> TS.valueAt xs lastIndex) bigSeries
         , bench "slice"  $ nf (\xs -> TS.valueAt (TS.slice xs (index 2) lastIndex) (index 1)) bigSeries
-        , bench "apply"  $ nf (TS.apply (+ 2)) bigSeries
+        , bench "fmap"  $ nf (fmap (+ 2)) bigSeries
         ]
     , bgroup "IO"
         [ bench "load CSV" $ nfIO (TS.loadCSV TS.NoHeader parseISODateTime "../testdata/test-100K.csv") ]