+++ src/Data/TimeSeries/Series.hs
@@ -11,8 +11,6 @@ Definition and basic operation on Series.
 module Data.TimeSeries.Series
     ( DataPoint
     , Series
-    , Value
-    , apply
     , emptySeries
     , series
     , size
@@ -20,7 +18,7 @@ module Data.TimeSeries.Series
     , toList
     , tsSeries
     , valueAt
-    , vfold
+    , values
     ) where
 
 import Prelude hiding (max, min)
@@ -28,24 +26,35 @@ import Data.Time (UTCTime)
 import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
 
 
-type Value = Double
 -- | Data points is single index value od time series
-data DataPoint = DP !UTCTime !Value
+data DataPoint a = DP !UTCTime a
                  deriving (Show, Eq)
+
+instance Functor DataPoint where
+    fmap f (DP x y) = DP x (f y)
+
+instance Foldable DataPoint where
+    foldMap f (DP _ y) = f y
 
 -- | Data structure for holding Series.
 -- Implementation should be hidden so it can be changed in the future
-data Series = Series [DataPoint]
+data Series a = Series [DataPoint a]
     deriving (Show, Eq)
+
+instance Functor Series where
+    fmap f (Series xs) = Series (map (fmap f) xs)
+
+instance Foldable Series where
+    foldMap f (Series xs) = foldMap (foldMap f) xs
 
 
 -- | Create empty series
-emptySeries :: Series
+emptySeries :: Series a
 emptySeries = Series []
 
 
--- | Create empty series
-series :: [(UTCTime, Double)] -> Series
+-- | Create series
+series :: [(UTCTime, a)] -> Series a
 series xs = Series $ map (\(x, y) -> DP x y) xs
 
 
@@ -53,14 +62,19 @@ series xs = Series $ map (\(x, y) -> DP x y) xs
 --
 -- >seriesFromSeconds [1, 2, 3] [41.3, 52.22, 3.0] == Series [DP 1970-01-01 00:00:01 UTC 2.3,DP 1970-01-01 00:00:02 UTC 4.5]
 --
-tsSeries :: [Integer] -> [Value] -> Series
+tsSeries :: [Integer] -> [a] -> Series a
 tsSeries ts vs = Series (zipWith DP idx vs)
     where idx = map (posixSecondsToUTCTime . fromIntegral) ts
 
 
 -- | Convert Time Series to list
-toList :: Series -> [(UTCTime, Double)]
+toList :: Series a -> [(UTCTime, a)]
 toList (Series xs) = map (\(DP x y) -> (x, y)) xs
+
+
+-- | Get only series values
+values :: Series a -> [a]
+values ts = map (\(_, y) -> y) (toList ts)
 
 
 -- | Get series size.
@@ -68,17 +82,12 @@ toList (Series xs) = map (\(DP x y) -> (x, y)) xs
 --
 -- >size (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) == 3
 --
-size :: Series -> Int
+size :: Series a -> Int
 size (Series xs) = length xs
 
 
--- | Apply function to every data point value
-apply :: (Value -> Value) -> Series -> Series
-apply f (Series xs) = Series $ map (\(DP x y) -> DP x (f y)) xs
-
-
 -- | Fold the values
-vfold :: (Double -> Value -> Double) -> Double -> Series -> Value
+vfold :: (b -> a -> b) -> b -> Series a -> b
 vfold f r (Series xs) = foldl (\d (DP _ y) -> f d y) r xs
 
 
@@ -88,10 +97,10 @@ vfold f r (Series xs) = foldl (\d (DP _ y) -> f d y) r xs
 -- >valueAt (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) 2 == Just 52.22
 -- >valueAt (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) 5 == Nothing
 --
-valueAt :: Series -> UTCTime -> Maybe Value
+valueAt :: Series a -> UTCTime -> Maybe a
 valueAt (Series xs) ts = safeHead [y | DP x y <- xs, x == ts]
     where safeHead [] = Nothing
-          safeHead (a:_) = Just a
+          safeHead (i:_) = Just i
 
 
 -- | Return series subset
@@ -100,7 +109,7 @@ valueAt (Series xs) ts = safeHead [y | DP x y <- xs, x == ts]
 -- >slice (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) 2 3 == Series [DP 2 52.22, DP 3 3.0]
 -- >slice (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) 5 23 == Series []
 --
-slice :: Series -> UTCTime -> UTCTime -> Series
+slice :: Series a -> UTCTime -> UTCTime -> Series a
 slice (Series xs) start end = Series [DP x y | DP x y <- xs, x >= start && x <= end]