+++ src/Data/TimeSeries/Stats.hs
@@ -5,7 +5,7 @@ module Data.TimeSeries.Stats
     ) where
 
 import Prelude hiding (max, min)
-import Data.TimeSeries.Series (Series, Value, vfold)
+import Data.TimeSeries.Series (Series)
 
 
 -- | Return maximum value in the series
@@ -14,9 +14,10 @@ import Data.TimeSeries.Series (Series, Value, vfold)
 -- >max (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) == 52.22
 -- >max (Series []) == 0
 --
-max :: Series -> Value
-max xs = vfold max' 0.0 xs
+max :: Real a => Series a -> a
+max xs = foldr max' minValue xs
     where max' a b = if a > b then a else b
+          minValue = fromIntegral (minBound :: Int)
 
 -- | Return maximum value in the series
 -- Complexity O(n)
@@ -24,7 +25,7 @@ max xs = vfold max' 0.0 xs
 -- >min (Series [DP 1 41.3, DP 2 52.22, DP 3 3.0]) == 3.0
 -- >min (Series []) == 0
 --
-min :: Series -> Value
-min xs = vfold min' maxValue xs
+min :: Real a => Series a -> a
+min xs = foldr min' maxValue xs
     where min' a b = if a < b then a else b
           maxValue = fromIntegral (maxBound :: Int)